django/requirements.txt
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
PyMySQL==1.0.2
cryptography==39.0.0
djangorestframework-simplejwt==5.2.2
